+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 
 ## 2026-04-23
+- `17:57` -- **Updated** `apps/desktopProbe/src/pages/settings.tsx`
 - `17:57` -- **Added** `apps/desktopProbe/packagers/household/com.first2apply.updater.plist`
 - `17:57` -- **Updated** `apps/desktopProbe/src/pages/settings.tsx`
 - `17:56` -- **Updated** `apps/desktopProbe/src/pages/settings.tsx`